^r Extract provider credentials, policy render, and path security from resolve_credential_sources.go, move-only (gofmt, vet, build, package tests pass; declaration set verified identical; internal restructure, no behavior change)
Step 2 of 2 for the authresolve god-file split: the remaining cohesive
groups move verbatim to resolve_provider_credentials.go,
resolve_policy_render.go, and resolve_path_security.go, leaving the entry
points in resolve_credential_sources.go.
